Fixed warnings
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@53976 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ web/src/main/java/org/geogebra/web/web/gui/toolbarpanel/ToolbarPanel.java
@@ -133,7 +133,8 @@ public class ToolbarPanel extends FlowPanel {
 		ToolbarDockPanelW dockPanel = getToolbarDockPanel();
 		DockSplitPaneW dockParent = dockPanel != null ? dockPanel.getParentSplitPane() : null;
 		if (dockPanel != null && lastOpenWidth != null) {
-			dockParent.setWidgetSize(dockPanel, header.isOpen() ? lastOpenWidth : CLOSED_WIDTH);
+			dockParent.setWidgetSize(dockPanel,
+					header.isOpen() ? lastOpenWidth.intValue() : CLOSED_WIDTH);
 			dockPanel.deferredOnResize();
 		}
